Version in branch just queries StackExchangeAPI for questions that meet criteria.  It was not clear if I was also to fetch the answers and display them in collection view cell also.
So I created a new branch getanswerswithapi to retrieve answers for questions and display them in collection view cell.  The added content is making the collection view stutter when scrolling which I assume from past experience is using constraints.  You could use frame positioning but with label using lines = 0 this would be tricky to implement.